# Conflicts:
#	cypress/integration/tests/Visualizer.spec.js
#	src/app/__consts__/InitialChartDataConst.js
#	src/app/components/charts/piechart/PieChart.jsx
#	src/app/mediators/ComponentMediators/PaneMediators/VisPaneMediator/VizPaneMediator.js
#	src/app/mediators/ModuleMediators/HomeModuleMediator/HomeModuleMediator.js
#	src/app/modules/UserManagement/CreateTeam/CreateTeamModule.jsx

+++ server/controllers/UserController.js
@@ -3,6 +3,8 @@ const general = require('./generalResponse');
 
 const User = require('../models/User');
 
+const isEqual = require('lodash/isEqual');
+
 const UserApi = {
   getUser: (req, res) => {
     const { authId } = req.query;
@@ -65,7 +67,7 @@ const UserApi = {
         avatar: user.avatar,
         firstName: user.firstName,
         lastName: user.lastName,
-        team: user.team
+        teams: user.teams
       },
       { new: true }
     )
@@ -95,8 +97,12 @@ const UserApi = {
           userFound.firstName = user.firstName;
         if (userFound.lastName !== user.lastName && user.lastName)
           userFound.lastName = user.lastName;
-        if (userFound.team !== user.team && user.team)
-          userFound.team = user.team;
+        if (userFound.role !== user.role && user.role) {
+          userFound.role = user.role;
+        }
+        if (!isEqual(userFound.teams, user.teams) && user.teams) {
+          userFound.teams = user.teams;
+        }
 
         userFound.save(error => {
           if (err) general.handleError(res, error);
@@ -109,6 +115,7 @@ const UserApi = {
 
   // updateUser by Admin, basically used to update role
   // and team
+  // TODO: change way of saving teams
   updateUserByAdmin: (req, res) => {
     const { user, updateUser } = req.body;
 
@@ -151,26 +158,71 @@ const UserApi = {
         // and then to each user we add the specified team
 
         updateUsers.forEach(updatU => {
-          User.findOne({ authId: updatU.authId }, (userErr, userFound) => {
-            if (userErr) return general.handleError(res, userErr);
-
-            // now if user is found we save the team
-            // but its not a problem if the user is not found
-            // because if the user is not found most likely they have not
-            // activated their account via email confirmation
-            // and their team and roles will be added, once they activate their
-            // email and login
-            if (userFound) {
-              userFound.team = team;
-
-              return userFound.save(saveError => {
-                if (saveError) general.handleError(res, saveError);
-              });
+          User.findOneAndUpdate(
+            { authId: updatU.authId },
+            { $push: { teams: team } },
+            (err, node) => {
+              if (err) console.log(err);
             }
-          });
+          );
         });
 
         return res.json({ message: 'user teams updated' });
+      }
+      return general.handleError(res, 'unauthorized', 401);
+    });
+  },
+
+  updateTeamAndUsersOfIt: (req, res) => {
+    const { user, team, usersToAdd, usersToDelete } = req.body;
+
+    // Find the current admin user
+    User.findOne({ authId: user.authId }, (adminErr, adminUser) => {
+      if (adminErr) return general.handleError(res, adminErr);
+      if (!adminUser) return general.handleError(res, 'User not found', 404);
+
+      if (adminUser.role === 'Administrator') {
+        // Remove team from selected users
+        if (usersToDelete.length > 0) {
+          usersToDelete.forEach(user => {
+            User.findOneAndUpdate(
+              { authId: user.authId },
+              { $pull: { teams: team.oldName } },
+              (err, node) => {
+                if (err) console.log(err);
+              }
+            );
+          });
+        }
+        // Add team to selected users
+        if (usersToAdd.length > 0) {
+          usersToAdd.forEach(user => {
+            User.findOneAndUpdate(
+              { authId: user.authId },
+              { $push: { teams: team.newName } },
+              (err, node) => {
+                if (err) console.log(err);
+              }
+            );
+          });
+        }
+
+        // Rename team and update it in all users that are in it
+        if (team.oldName !== team.newName) {
+          User.updateMany(
+            { teams: { $eq: team.oldName } },
+            { $set: { 'teams.$[element]': team.newName } },
+            {
+              upsert: false,
+              arrayFilters: [{ element: { $eq: team.oldName } }]
+            },
+            (err, node) => {
+              if (err) console.log(err);
+            }
+          );
+        }
+
+        return res.json({ message: 'users & team updated' });
       }
       return general.handleError(res, 'unauthorized', 401);
     });
@@ -189,14 +241,37 @@ const UserApi = {
   // },
 
   deleteUser: (req, res) => {
-    const { delId } = req.query;
-    // TODO: should be adjusted without the promises, or maybe with promises if
-    // TODO: it works and makes sense
-    User.deleteOne({ authId: delId }, error => {
-      if (error) general.handleError(res, error);
+    const { delId, authId } = req.query;
 
-      return res.json({ message: 'user deleted' });
+    User.findOne({ authId }, (adminErr, adminUser) => {
+      if (adminErr) general.handleError(res, adminErr);
+      else if (!adminUser)
+        general.handleError(res, 'Admin user not found', 404);
+      else if (
+        adminUser.role === 'Administrator' ||
+        adminUser.role === 'Super admin'
+      )
+        User.deleteOne({ authId: delId }, error => {
+          if (error) general.handleError(res, error);
+          else res.json({ message: 'user deleted' });
+        });
+      else {
+        general.handleError(res, 'Unauthorized', 401);
+      }
     });
+  },
+
+  deleteTeam: (req, res) => {
+    const { name } = req.body;
+
+    User.updateMany(
+      { teams: { $eq: name } },
+      { $pull: { teams: name } },
+      (err, node) => {
+        if (err) return general.handleError(res, err);
+      }
+    );
+    return res.json({ message: 'users & team updated' });
   }
 };
